LayoutTests/accessibility/gtk/text-for-range-embedded-objects-expected.txt

Choose: foo bar (pick one)

Choose:  (pick one)



This tests the ability to get element text for a specified range.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS paragraph1.stringForRange(0, -1) is 'Choose: ￼foo ￼bar (pick one)'
PASS paragraph2.stringForRange(0, -1) is 'Choose: ￼ (pick one)'
PASS paragraph3.stringForRange(0, -1) is '￼'
PASS successfullyParsed is true

TEST COMPLETE
